# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Oxidean REST API
  version: '1'
  description: Public REST API for Oxidean — a companion to the typed JSON RPC at `POST /api/rpc`. Each route dispatches to the same internal procedure the RPC layer uses, so ACLs, PAT scope gates, and side effects are identical. Generated from the route table — this document cannot drift from the served surface.
servers:
- url: /api/v1
security:
- sessionCookie: []
- bearerPat: []
...
paths:
...
  /orgs:
    post:
      tags:
      - orgs
      summary: Create an org (`org.create`) — session cookie only
      operationId: createOrg
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - slug
              title: CreateOrgRequest
              description: '`org.create` input — slug + optional display name (plan 10-13).'
              type: object
              properties:
                slug:
                  type: string
                display_name:
                  type:
                  - string
                  - 'null'
                  default: null
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{slug}:
    get:
      tags:
      - orgs
      summary: Get an org (`org.get`) — anonymous OK for public org data
      operationId: getOrg
      security: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{slug}/members:
    get:
      tags:
      - orgs
      summary: Org members (`org.members.list`)
      operationId: listOrgMembers
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgMembersListResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orgs/{slug}/repos:
    get:
      tags:
      - orgs
      - repos
      summary: Repos under an org (`repo.listByOwner`)
      operationId: listOrgRepos
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoListMineResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
      - orgs
      - repos
      summary: Create an org repo (`repo.create` with `owner` = slug)
      operationId: createOrgRepo
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
              - name
              title: CreateRepoRequest
              description: 'Create-repository input (RPC wired in 07-12; templates in 07-03 / issue #18).'
              type: object
              properties:
                name:
                  type: string
                description:
                  type:
                  - string
                  - 'null'
                  default: null
                visibility:
                  description: When omitted, API uses instance `default_visibility` (else public) — D-08.
                  anyOf:
                  - $ref: '#/components/schemas/RepoVisibility'
                  - type: 'null'
                  default: null
                stack_id:
                  description: |-
                    Stack preset pack id under `assets/stack-presets/` (omit / null = none).
                    Mutually exclusive with `instance_pack_id` / `template_repo_id`.
                  type:
                  - string
                  - 'null'
                  default: null
                instance_pack_id:
                  description: |-
                    Instance-admin template pack id (omit / null = none). Mutually exclusive
                    with `stack_id` / `template_repo_id`.
                  type:
                  - string
                  - 'null'
                  default: null
                template_repo_id:
                  description: |-
                    Source template repository id (omit / null = none). Mutually exclusive
                    with `stack_id` / `instance_pack_id`.
                  type:
                  - string
                  - 'null'
                  default: null
                license_id:
                  description: SPDX license id or omit / null / `"none"` for no LICENSE file.
                  type:
                  - string
                  - 'null'
                  default: null
                gitignore_id:
                  description: Gitignore catalog id under `assets/gitignore/` (omit / null / `"none"` = none).
                  type:
                  - string
                  - 'null'
                  default: null
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepoPublic'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/users:
    get:
      tags:
      - admin
      summary: List users (`admin.users.list`; sys-admin session only)
      operationId: adminListUsers
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type:
          - string
          - 'null'
          default: null
      - name: limit
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          default: null
      - name: offset
        in: query
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
          default: null
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersListResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /admin/lfs/usage:
    get:
      tags:
      - admin
      summary: Instance LFS usage stats (`admin.lfs.getUsage`; sys-admin session only)
      operationId: adminLfsUsage
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLfsUsageResponse'
        default:
          description: Error — `{ code, message, data? }` with a mapped HTTP status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: domain.error_code
        message:
          type: string
        data:
          description: optional structured detail
    RepoVisibility:
      description: 'Repo visibility. Serialized lowercase: `public` | `private`.'
      type: string
      enum:
      - public
      - private
...
    Role:
      type: string
      enum:
      - user
      - admin
      - sys-admin
      description: 'Instance-level user role. Serialized as kebab-case: `user` | `admin` | `sys-admin`.'
    RepoListMineResponse:
      type: object
      properties:
        repos:
          type: array
          items:
            $ref: '#/components/schemas/RepoPublic'
      required:
      - repos
      description: '`repo.listMine` — caller''s non-deleted repos, recently updated first (GIT-01 / D-13).'
    RepoPublic:
      type: object
      properties:
        id:
          type: string
        owner_id:
          type: string
        owner_type:
          $ref: '#/components/schemas/OwnerType'
          description: 'Polymorphic owner: `user` | `org` (D-ORG-01).'
        owner_username:
          type: string
          description: Public slug label (username or org slug).
        name:
          type: string
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/RepoVisibility'
        default_branch:
          type: string
        updated_at:
          type: string
        can_admin:
          type: boolean
          description: Caller has Admin capability (D-ORG-05 / settings UI).
          default: false
        can_write:
          type: boolean
          description: Caller has Write capability (D-ORG-05).
          default: false
        star_count:
          type: integer
          format: int64
          description: Denormalized star counter (D-SOC-02 / D-SOC-03).
          default: 0
        open_issue_count:
          type: integer
          format: int64
          description: Open issue count for the Issues tab badge (populated by repo.get).
          default: 0
        open_pull_count:
          type: integer
          format: int64
          description: Open pull-request count for the Pulls tab badge (populated by repo.get).
          default: 0
        viewer_has_starred:
          type: boolean
          description: Whether the authenticated viewer has starred this repo.
          default: false
        is_fork:
          type: boolean
          description: True when this repository is a fork of another.
          default: false
        is_template:
          type: boolean
          description: 'True when owners expose this repo as a create-from template (issue #18).'
          default: false
        archived:
          type: boolean
          description: |-
            Read-only archive mode (GIT-20): browse/clone stay open; pushes and
            content writes (issues, pulls, releases, branches, LFS uploads) are blocked.
          default: false
        homepage:
          type: string
          description: 'Project homepage URL / text (issue #23).'
          default: ''
        topics:
          type: array
          items:
            type: string
          description: 'Topic slugs (issue #23).'
          default: []
        fork_count:
          type: integer
          format: int64
          description: 'Active forks in this repo''s network (issue #23).'
          default: 0
        watch_count:
          type: integer
          format: int64
          description: 'Watch / subscribe counter (issue #23).'
          default: 0
        viewer_is_watching:
          type: boolean
          description: |-
            Whether the authenticated viewer is watching this repo
            (subscription row at a non-`ignore` level).
          default: false
        viewer_watch_level:
          anyOf:
          - $ref: '#/components/schemas/WatchLevel'
          - type: 'null'
          description: |-
            Authenticated viewer's watch level — `all` | `participating` | `ignore`;
            absent when the viewer has no subscription row (DEBT-06).
        issues_enabled:
          type: boolean
          description: |-
            Issues unit enabled — gates `issue.*` RPCs + the Issues tab (COL-13).
            Populated by `repo.get`; list endpoints may report the default `true`.
          default: true
        pulls_enabled:
          type: boolean
          description: |-
            Pull-requests unit enabled — gates `pull.*` RPCs + the Pulls tab (COL-13).
            Populated by `repo.get`; list endpoints may report the default `true`.
          default: true
        fork_network_id:
          type:
          - string
          - 'null'
          description: Fork network root id (own id for roots) — D-SOC-14 / D-PR-01.
        forked_from:
          anyOf:
          - $ref: '#/components/schemas/ForkParentSummary'
          - type: 'null'
          description: Immediate parent when `is_fork` (D-SOC-16).
      required:
      - id
      - owner_id
      - owner_type
      - owner_username
      - name
      - description
      - visibility
      - default_branch
      - updated_at
      description: Public repository metadata returned over RPC.
    OwnerType:
      type: string
      enum:
      - user
      - org
      description: |-
        Polymorphic repository owner discriminant (D-ORG-01).
        Serialized lowercase: `user` | `org`.
    WatchLevel:
      oneOf:
      - type: string
        const: all
        description: Notified on all repo activity (default for `repo.watch`).
      - type: string
        const: participating
        description: Notified only when participating or @-mentioned (stored default).
      - type: string
        const: ignore
        description: Never notified for this repo — suppresses participation + mention rows.
      description: |-
        Per-user repository watch level — the notification matrix (DEBT-06).
        Serialized snake_case: `all` | `participating` | `ignore`.
    ForkParentSummary:
      type: object
      properties:
        id:
          type: string
        owner:
          type: string
        name:
          type: string
      required:
      - id
      - owner
      - name
      description: Parent summary when this repo is a fork (D-SOC-16).
...
    MemberBasePermission:
      type: string
      enum:
      - none
      - read
      - write
      description: |-
        Default Member permission on org private repos (D-ORG-02b).
        Serialized lowercase: `none` | `read` | `write`.
    OrgRole:
      type: string
      enum:
      - owner
      - admin
      - member
      description: 'Org membership role (D-ORG-02a). Serialized lowercase: `owner` | `admin` | `member`.'
...
    OrgPublic:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        display_name:
          type: string
        member_base_permission:
          $ref: '#/components/schemas/MemberBasePermission'
        created_at:
          type: string
        updated_at:
          type: string
      required:
      - id
      - slug
      - display_name
      - member_base_permission
      - created_at
      - updated_at
      description: Public organization profile returned over RPC (ORG-01).
    OrgMembersListResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: '#/components/schemas/OrgMemberPublic'
      required:
      - members
      description: '`org.members.list` response.'
    OrgMemberPublic:
      type: object
      properties:
        user_id:
          type: string
        username:
          type: string
        role:
          $ref: '#/components/schemas/OrgRole'
        created_at:
          type: string
      required:
      - user_id
      - username
      - role
      - created_at
      description: Public membership row — no email (ORG-01).
...
    AdminUsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/AdminUserPublic'
        total:
          type: integer
          format: int64
      required:
      - users
      - total
      description: '`admin.users.list` response.'
    AdminUserPublic:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        display_name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        email_verified:
          type: boolean
        banned_at:
          type:
          - string
          - 'null'
        created_at:
          type: string
      required:
      - id
      - email
      - username
      - display_name
      - role
      - email_verified
      - created_at
      description: Admin user row for `admin.users.list` (includes email + ban state).
    AdminLfsUsageResponse:
      type: object
      properties:
        physical_bytes:
          type: integer
          format: int64
        object_count:
          type: integer
          format: int64
        logical_bytes:
          type: integer
          format: int64
        by_repo:
          type: array
          items:
            $ref: '#/components/schemas/AdminLfsRepoUsageEntry'
        by_owner:
          type: array
          items:
            $ref: '#/components/schemas/AdminLfsOwnerUsageEntry'
      required:
      - physical_bytes
      - object_count
      - logical_bytes
      - by_repo
      - by_owner
      description: '`admin.lfs.getUsage` — instance breakdown (D-LFS-19).'
    AdminLfsRepoUsageEntry:
      type: object
      properties:
        repository_id:
          type: string
        owner:
          type: string
        name:
          type: string
        object_count:
          type: integer
          format: int64
        logical_bytes:
          type: integer
          format: int64
      required:
      - repository_id
      - owner
      - name
      - object_count
      - logical_bytes
      description: Per-repo row in admin instance usage breakdown.
    AdminLfsOwnerUsageEntry:
      type: object
      properties:
        owner_id:
          type: string
        owner_slug:
          type: string
        object_count:
          type: integer
          format: int64
        logical_bytes:
          type: integer
          format: int64
      required:
      - owner_id
      - owner_slug
      - object_count
      - logical_bytes
      description: Per-owner (user/org) row in admin instance usage breakdown.
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: oxidean_session
      description: Browser session cookie — wins if both credentials are sent.
    bearerPat:
      type: http
      scheme: bearer
      description: Personal access token. Classic `repo` covers the repo domain; `admin.*` and org/credential mutations are session-only.